refactor(config): extract audit-template helpers from operations.audit
